feature: fetch the most current name for HSPs
Previously, we would send the user the display_name that was used when
the profile was recorded. Instead, we now fetch the current display_name
and use that.

There will be a follow-up PR to not record the display_name in the
first place, since it's now a dynamically created field.