feature: use captured_date instead of created_on
Previously, we displayed the "created" date of the historical record.
This made little sense for end users since they are not interested in
when a record was written to our DB.

Instead, we display "captured_date" which is when the data was
captured on the host system. If the captured date is not found, we
fall back to using the created date for that field.