add expiration env var
This commit adds a new config env var `VALID_PROFILE_AGE_DAYS`, which
defaults to 7 days. Profiles with a captured_date older than this will
not be returned in the list of profiles for a system.  Additionally,
if a profile is older than the date, it will not be returned when
fetched directly via API.

It is possible to supply fractional days (0.01, 3.5, etc).