fix: roll back ORM transaction on exception
Previously, if the database connection died while message processing
was happening, we would keep attempting to re-use the dead connection
instead of creating a new one. This would occur even after the
database came back up.

We now roll back the ORM transaction if there is an error. This lets
us try with a fresh connection on the next run, so if the DB is back
up we will reconnect.