add index on inventory_id
When we added the check for duplicate IDs + captured_dates, we ended
up throwing a lot of seqential scans at the database. For example, a
single query for a particular ID + date took around 500 msec.

This commit adds an index on inventory_id. After adding, a query in my
environment went from 5.475 ms to 0.042 ms.