fix: build with psycopg2 from source
Inside the govcloud environment, all apps must run with fips compliance.
The psycopg2-binary does not contain what it needs to run in a FIPS
environment. To fix this, we need to compile psycopg2 from source when
we build the image.

This was done wiht host-inventory successfully, so we're going through
other apps with a similar setup and fixing the image to run properly.
This shouldn't effect the way the app runs in other environments.

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>